requirements.txt
bs4==0.0.1
fastapi==0.74.1
lxml==4.7.1
mysql-connector-python==8.0.28
pre-commit==2.17.0
python-dotenv==0.19.2
requests==2.27.1
uvicorn==0.17.5
